Fixed Juice Up Banner on Index
Fixed the Juice Up Banner on the Index page so that there isn't weird empty space around the text on some viewports. Also need to note that the JS file for the juice up banner may not be needed anymore as it make the banner grow or shrink in relation to the viewport width, but now with the height set to auto, this effect seems to be minimal or none in my testing environment.