Updated JSON Database
Removed duplicate option. However do note that there is a A1278 MacBook Pro, and an A1278 MacBook. Probably not like this from Apple, but all models were copied from MobileSentrix layout. If price for battery becomes 2 separate prices update to make a distinction in the code between the two models.